implement remaining in vehicle on teleporting
This was painful to get work. It seems to abort the teleport when
there is an entity in the way. But only when inside a boat. Therefore
I had to check for an entity in the way and this don't work when the chunk isn't loaded
as world.isSpaceEmpty returns true in this case regardless if there will be an entity
when completely loaded